fix scalar index creation
