+++ client/src/utils/search.ts
@@ -1,11 +1,7 @@
 import { Field } from '../components/advancedSearch/Types';
 import { DataTypeEnum, DataTypeStringEnum } from '../pages/collections/Types';
-import {
-  FieldData,
-  IndexType,
-  IndexView,
-  INDEX_TYPES_ENUM,
-} from '../pages/schema/Types';
+import { FieldData, IndexType, IndexView } from '../pages/schema/Types';
+import { INDEX_TYPES_ENUM } from '@/consts';
 import { FieldOption } from '../types/SearchTypes';
 
 /**
